v2

+++ src/controller/client.controller.js
@@ -1,7 +1,10 @@
-import clientService from '../service/sales.service.js';
+import clientService from '../service/client.service.js';
 
 async function novoClient(req, res, next) {
     try {
+        
+        if(req.userClass != 'admin') throw new Error('vc nao tem permição para cadastrar clientes');
+
         let client = req.body;
         if (!client.nome || !client.endereco || !client.cidade || !client.sexo) {
             throw new Error('todas as informacoes sao necessárias para cadastrar o cliente')
@@ -9,6 +12,7 @@ async function novoClient(req, res, next) {
         client = await clientService.novoClient(client);
         res.send(client);
         logger.info(`POST /client - ${JSON.stringify(client)}`);
+
     } catch (error) {
         next(error)
     }
@@ -16,6 +20,7 @@ async function novoClient(req, res, next) {
 
 async function getClients(req, res, next) {
     try {
+        if(req.userClass != 'admin') throw new Error('vc nao tem permição para consultar clientes');
         res.send(await clientService.getClients())
         logger.info(`GET /client`);
     } catch (error) {
@@ -25,7 +30,10 @@ async function getClients(req, res, next) {
 
 async function getClient(req, res, next) {
     try {
-        res.send(await clientService.getClient(req.params.id))
+        if(req.userClass != 'admin') throw new Error('vc nao tem permição para consultar clientes');
+        let client = await clientService.getClient(req);
+        if(!client) return res.status(404).send('cliente nao existe');
+        res.send(client);
         logger.info(`GET /client:id`);
     } catch (error) {
         next(error)
@@ -34,6 +42,11 @@ async function getClient(req, res, next) {
 
 async function deleteClient(req, res, next) {
     try {
+
+        if(req.userClass != 'admin') throw new Error('vc nao tem permição para deletar clientes');
+        let client = await clientService.getClient(req);
+        if(!client) return res.status(404).send('cliente nao existe');
+
         await clientService.deleteClient(req.params.id);
         res.end();
         logger.info(`DELETE /client/${req.params.id} `)
@@ -44,7 +57,13 @@ async function deleteClient(req, res, next) {
 
 async function updateClient(req, res, next) {
     try {
-        let client = req.body;
+
+        if(req.userClass != 'admin') throw new Error('vc nao tem permição para atualizar clientes');
+        req.params.id = req.body.client_id;
+        let client = await clientService.getClient(req);
+        if(!client) return res.status(404).send('cliente nao existe');
+
+        client = req.body;
         if (!client.nome || !client.endereco || !client.cidade || !client.sexo || !client.client_id) {
             throw new Error('todas as informacoes sao necessárias para atualizar o cliente')
         }
